Got the api method for the last feed working. Changed some tests, as the data is not reliable.

+++ Stockfighter.Client/Data/Balances.cs
@@ -5,6 +5,6 @@
     /// </summary>
     public class Balances
     {
-        public decimal Usd { get; set; }
+        public int Usd { get; set; }
     }
 }
